+++ src/components/Avatar/Avatar.tsx
@@ -7,20 +7,16 @@ import Empty from '../../assets/static/empty.svg';
 
 export interface IAvatarProps {
     src?: string;
-    size?: number;
+    size?: number | string;
     defaultSrc?: string;
     type?: 'circle' | 'square';
 }
 
-export const Avatar = ({
-    src = Empty,
-    size = 38,
-    type = 'square',
-}: IAvatarProps) => (
+export const Avatar = ({ src, size = 38, type = 'square' }: IAvatarProps) => (
     <img
         width={size}
         height={size}
         className={`unique-avatar ${type}`}
-        src={src}
+        src={src?.length ? src : Empty}
     />
 );