+++ src/components/Link/Link.tsx
@@ -6,13 +6,18 @@ import { AnchorHTMLAttributes } from 'react';
 import './Link.scss';
 
 export interface LinkProps extends AnchorHTMLAttributes<HTMLAnchorElement> {
-    title: string;
+    title?: string;
     role?: 'primary' | 'secondary' | 'danger';
     onClick?(): void;
 }
 
-export const Link = ({ title, role = 'primary', ...rest }: LinkProps) => (
+export const Link = ({
+    children,
+    title,
+    role = 'primary',
+    ...rest
+}: LinkProps) => (
     <a className={`unique-link ${role}`} {...rest}>
-        {title}
+        {children || title}
     </a>
 );