+++ src/stories/Link.stories.tsx
@@ -1,9 +1,9 @@
 import { ComponentStory, ComponentMeta } from '@storybook/react';
-import { Link } from '../components';
+import { Icon, Link } from '../components';
 
 export default {
     title: 'Components/Link',
-    component: Link
+    component: Link,
 } as ComponentMeta<typeof Link>;
 
 const Template: ComponentStory<typeof Link> = (args) => <Link {...args} />;
@@ -11,7 +11,7 @@ const Template: ComponentStory<typeof Link> = (args) => <Link {...args} />;
 export const Primary = Template.bind({});
 
 Primary.args = {
-    title: 'Primary link'
+    title: 'Primary link',
 };
 
 Primary.storyName = 'Primary';
@@ -20,7 +20,7 @@ export const Secondary = Template.bind({});
 
 Secondary.args = {
     title: 'Secondary link',
-    role: 'secondary'
+    role: 'secondary',
 };
 
 Secondary.storyName = 'Secondary';
@@ -29,7 +29,7 @@ export const Danger = Template.bind({});
 
 Danger.args = {
     title: 'Danger link',
-    role: 'danger'
+    role: 'danger',
 };
 
 Danger.storyName = 'Danger';
@@ -38,7 +38,24 @@ export const Disabled = Template.bind({});
 
 Disabled.args = {
     title: 'Disabled link',
-    role: 'disabled'
+    role: 'disabled',
 };
 
 Disabled.storyName = 'Disabled';
+
+const TemplateNode: ComponentStory<typeof Link> = (args) => (
+    <Link>
+        External link
+        <Icon
+            size={18}
+            color="var(--color-primary-500)"
+            name="arrow-up-right"
+        />
+    </Link>
+);
+
+export const PrimaryNode = TemplateNode.bind({});
+
+PrimaryNode.args = {};
+
+PrimaryNode.storyName = 'Primary w/ children';