+++ src/widgets/AccountsManager/components/AccountsManagerDropdown/AccountsManagerDropdown.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { Dropdown, Icon, Link, Text, Toggle } from '../../../../components';
 import { AccountsManagerProps } from '../../AccountsManager';
 import { AccountCard } from '../index';
@@ -15,12 +15,18 @@ export const AccountsManagerDropdown = ({
     manageBalanceLinkTitle,
     balance,
     symbol,
+    isTouch,
     avatarRender,
     onAccountChange,
     onNetworkChange,
     onManageBalanceClick,
+    onCopyAddressClick,
 }: AccountsManagerProps) => {
     const [copied, copy] = useCopyToClipboard();
+
+    useEffect(() => {
+        copied && onCopyAddressClick?.(copied);
+    }, [copied]);
 
     return (
         <div className="accounts-manager-dropdown">
@@ -43,6 +49,7 @@ export const AccountsManagerDropdown = ({
                         size: 8,
                     }}
                     onChange={onAccountChange}
+                    isTouch={isTouch}
                 >
                     <div
                         className="accounts-select"