+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # CHANGELOG
 
+## 0.0.35 August 19, 2022
+
+Changes:
+
+-   AccountsManager widget responsive add, onCopyAddressClick handler add
+-   Upload component validation add
+-   Notification component context memo add
+-   Link component children add
+-   Avatar component size prop fix, src empty string fix
+-   Pagination component static text fix
+
 ## 0.0.34 August 12, 2022
 
 Changes: